Testando correção no Jenkinsfile

+++ Jenkinsfile
@@ -84,12 +84,12 @@ pipeline {
         always {
             script {
                 // Preservar os containers antigos enquanto os novos são iniciados
-                def oldContainers = sh(script: 'docker ps -a -q --filter "name=spring-petclinic"', returnStdout: true).trim().tokenize('\\n')
+                def oldContainers = sh(script: 'docker ps -a -q --filter "name=spring-petclinic"', returnStdout: true).trim().tokenize('\n')
                 def newProdContainerId = sh(script: 'docker ps -q --filter "name=spring-petclinic-prod"', returnStdout: true).trim()
 
                 // Remover os containers antigos parados 
                 oldContainers.each { containerId ->
-                    if (containerId != newProdContainerId) {
+                    if (containerId && containerId != newProdContainerId) {
                         sh "docker rm -f ${containerId}"
                     }
                 }
